components/deployment-model-evaluation/requirements.txt
pandas==1.5.3
scikit-learn==1.2.2
joblib==1.3.2
google-cloud-storage==2.10.0
numpy==1.23.5
kpf 